+++ test/node_test.dart
@@ -9,8 +9,6 @@ import 'dart:io';
 import 'package:supply_chain/supply_chain.dart';
 import 'package:test/test.dart';
 
-import 'shared_tests.dart';
-
 void main() {
   late Scm scm;
   late Scope chain;
@@ -900,7 +898,224 @@ void main() {
     });
 
     group('smartNodes', () {
-      smartNodeTest();
+      group('a node should become a smart node', () {
+        test('when it is placed within a smart scope', () {
+          // Create a root scope
+          final scope = Scope.example();
+
+          // Create a master scope within the root scope
+          // containing one node.
+          scope.mockContent({
+            'master': {
+              'node': 0,
+            },
+          });
+
+          // Create a non smart node called follower
+          const followerNodeBp = NodeBluePrint<int>(
+            key: 'node',
+            initialProduct: 0,
+          );
+          expect(followerNodeBp.isSmartNode, isFalse);
+
+          // Create a follower scope
+          final followerScope = const ScopeBluePrint(
+            key: 'follower',
+            smartMaster: ['master'],
+          ).instantiate(hostScope: scope);
+
+          // The follower scope is a smart scope
+          expect(followerScope.isSmartScope, isTrue);
+
+          // Instantiate the followerNode blue print within the
+          // follower smart scope
+          final followerNode = followerNodeBp.instantiate(scope: followerScope);
+
+          // Although the follower node's blue print is not a smart node,
+          // the follower node is a smart node, because it is placed within
+          // a smart scope.
+          expect(followerNode.isSmartNode, isTrue);
+
+          // The right smart master path is assigned to the node
+          expect(followerNode.smartMaster, ['master', 'node']);
+
+          // Therefore changing the master node's value
+          // should change the follower node's
+          expect(followerNode.product, 0);
+          final masterNode = scope.findNode<int>('master.node')!;
+          masterNode.product = 1;
+          scope.scm.testFlushTasks();
+          expect(followerNode.product, 1);
+
+          // Anyway, meta nodes should not be smart nodes
+          bool didCheck = false;
+          for (final metaScope in followerScope.metaScopes) {
+            for (final metaNode in metaScope.nodes) {
+              expect(metaNode.isSmartNode, isFalse);
+              didCheck = true;
+            }
+          }
+          expect(didCheck, isTrue);
+        });
+      });
+
+      test('should work', () {
+        var smartNodeValue = 2;
+        var master0Value = 3;
+        var master1Value = 4;
+
+        final scope = Scope.example();
+        final scm = scope.scm;
+        scope.mockContent({
+          'a': {
+            'b': {
+              'c': {
+                'height': NodeBluePrint<int>(
+                  key: 'height',
+                  initialProduct: smartNodeValue,
+                  smartMaster: ['x', 'height'],
+                ),
+                'd': {
+                  'customer': NodeBluePrint<int>.map(
+                    supplier: 'height',
+                    initialProduct: 0,
+                    toKey: 'customer',
+                  ),
+                },
+              },
+            },
+          },
+        });
+
+        final smartNode = scope.findNode<int>('c.height')!;
+        expect(smartNode.bluePrint.isSmartNode, true);
+        final a = scope.findScope('a')!;
+        final b = scope.findScope('b')!;
+        final customer = scope.findNode<int>('d.customer')!;
+        scm.testFlushTasks();
+
+        // ..............................................
+        // Use smartNode itself when no smartNode is available
+
+        // SmartNode delivers it's own initial value
+        // because no other master height node can be found
+        expect(smartNode.product, smartNodeValue);
+
+        // The customer uses the place holder
+        expect(customer.product, smartNodeValue);
+
+        // .........................................
+        // Add master0 replacing the smartNode
+
+        // Add x.height to the scope a
+        a.mockContent({
+          'x': {
+            'height': master0Value,
+          },
+        });
+        final master0 = scope.findNode<int>('x.height')!;
+        scm.testFlushTasks();
+
+        // Now master0 should deliver the value of the smartNode
+        expect(master0.product, master0Value);
+        expect(smartNode.product, master0Value);
+        expect(customer.product, master0Value);
+
+        // Change the master0
+        // SmartNode value should be updated
+        master0Value *= 10;
+        master0.product = master0Value;
+        scm.testFlushTasks();
+
+        expect(master0.product, master0Value);
+        expect(smartNode.product, master0Value);
+        expect(customer.product, master0Value);
+
+        // ..........................................................
+        // Insert another master1 between smartNode and master0
+        b.mockContent({
+          'x': {
+            'height': master1Value,
+          },
+        });
+        final master1 = scope.findNode<int>('b.x.height')!;
+        scm.testFlushTasks();
+
+        // Now the smartNode should deliver the value of the new smartNode
+        expect(master0.product, master0Value);
+        expect(master1.product, master1Value);
+        expect(smartNode.product, master1Value);
+        expect(customer.product, master1Value);
+
+        // .........................................................
+        // Remove the master1 between  smartNode and master0
+        master1.dispose();
+        scm.testFlushTasks();
+
+        // Now master0 should take over again
+        expect(master0.product, master0Value);
+        expect(smartNode.product, master0Value);
+        expect(customer.product, master0Value);
+
+        // .......................
+        // Remove the master0.
+        // SmartNode should take over again
+        master0.dispose();
+        scm.testFlushTasks();
+
+        expect(smartNode.suppliers, isEmpty);
+        smartNode.product = smartNodeValue;
+        expect(smartNode.product, smartNodeValue);
+        expect(customer.product, smartNodeValue);
+      });
+
+      test('should be able to connect to a master in own scope', () {
+        final scope = Scope.example();
+        final scm = scope.scm;
+        scope.mockContent({
+          'a': 5,
+          'b': const NodeBluePrint(
+            key: 'b',
+            initialProduct: 1,
+            smartMaster: ['a'],
+          ),
+        });
+        scm.testFlushTasks();
+        expect(scope.findNode<int>('b')!.product, 5);
+      });
+
+      test('should remove suppliers from disposed smart nodes', () {
+        final scope = Scope.example();
+        final scm = scope.scm;
+
+        // Create two sibling nodes that might reference each other.
+        scope.mockContent(
+          {
+            'scope0': {
+              'master': 0,
+              'a': {
+                'smartNode': const NodeBluePrint(
+                  key: 'smartNode',
+                  smartMaster: ['master'],
+                  initialProduct: 1,
+                ),
+              },
+            },
+          },
+        );
+
+        scm.testFlushTasks();
+
+        final master = scope.findNode<int>('master')!;
+        final smartNode = scope.findNode<int>('smartNode')!;
+        expect(smartNode.suppliers, [master]);
+
+        // Dispose smart node
+        smartNode.dispose();
+
+        // The smart node should not have any suppliers anymore.
+        expect(smartNode.suppliers, isEmpty);
+      });
     });
 
     group('initSuppliers', () {
@@ -965,7 +1180,7 @@ void main() {
         // Find the first master node in the hierarchy
         final nodeB = scope.findNode<int>('a.b.node')!;
         final nodeC = scope.findNode<int>('c.node')!;
-        final masterOfNodeC = nodeC.findMasterNode();
+        final masterOfNodeC = nodeC.findSmartMaster();
         expect(masterOfNodeC, nodeB);
 
         // Dispose the master node
@@ -974,7 +1189,7 @@ void main() {
 
         // The next master node should be found
         final nodeA = scope.findNode<int>('a.node')!;
-        final masterOfNodeA = nodeC.findMasterNode();
+        final masterOfNodeA = nodeC.findSmartMaster();
         expect(masterOfNodeA, nodeA);
 
         // Dispose the master node
@@ -982,7 +1197,7 @@ void main() {
         scope.scm.testFlushTasks();
 
         // No master node is found
-        final masterOfNodeX = nodeC.findMasterNode();
+        final masterOfNodeX = nodeC.findSmartMaster();
         expect(masterOfNodeX, isNull);
       });
 
@@ -1012,17 +1227,17 @@ void main() {
 
         scope.scm.testFlushTasks();
         final nodeB = scope.findNode<int>('b.node')!;
-        final masterNodeB = nodeB.findMasterNode();
+        final masterNodeB = nodeB.findSmartMaster();
         expect(masterNodeB, masterNode);
 
         final nodeC = scope.findNode<int>('c.node')!;
-        final masterNodeOfC = nodeC.findMasterNode();
+        final masterNodeOfC = nodeC.findSmartMaster();
         expect(masterNodeOfC, masterNode);
 
         masterNode.dispose();
         scope.scm.testFlushTasks();
-        expect(nodeB.findMasterNode(), isNull);
-        expect(nodeC.findMasterNode(), isNull);
+        expect(nodeB.findSmartMaster(), isNull);
+        expect(nodeC.findSmartMaster(), isNull);
       });
     });